included concatgen.h in concatgen.c and fixed return type

+++ concat/libavformat/concatgen.c
@@ -29,7 +29,7 @@
  *  process.
  */
 
-#include "playlist.h"
+#include "concatgen.h"
 
 int ff_concatgen_read_packet(AVFormatContext *s,
                              AVPacket *pkt)
@@ -90,10 +90,10 @@ int ff_concatgen_read_seek(AVFormatContext *s,
     return ic->iformat->read_seek(ic, stream_index, pts, flags);
 }
 
-int ff_concatgen_read_timestamp(AVFormatContext *s,
-                                int stream_index,
-                                int64_t *pos,
-                                int64_t pos_limit)
+int64_t ff_concatgen_read_timestamp(AVFormatContext *s,
+                                    int stream_index,
+                                    int64_t *pos,
+                                    int64_t pos_limit)
 {
     PlaylistContext *ctx;
     AVFormatContext *ic;
